+++ dessia_common/workflow/core.py
@@ -1900,20 +1900,22 @@ class WorkflowRun(WorkflowState):
 
         Concatenate WorkflowState display_settings and inserting Workflow ones.
         """
-        workflow_settings = self.workflow.display_settings()
+        workflow_settings = [display_setting for display_setting in self.workflow.display_settings()
+                             if display_setting.selector != "Documentation"]
         block_settings = self.workflow.blocks_display_settings
-        displays_by_default = [s.load_by_default for s in block_settings]
 
-        workflow_settings_to_keep = []
+        displays_by_default = [s.load_by_default for s in block_settings]
+        documentation = DisplaySetting(selector="Documentation", type_="markdown", method="to_markdown",
+                                       load_by_default=True)
+        documentation.load_by_default = not any(displays_by_default)
+
+        workflow_settings_to_keep = [documentation]
         for settings in workflow_settings:
             # Update workflow settings
             settings.compose("workflow")
 
             if settings.selector == "Workflow":
                 settings.load_by_default = False
-
-            if settings.selector == "Documentation":
-                settings.load_by_default = not any(displays_by_default)
 
             if settings.selector != "Tasks":
                 workflow_settings_to_keep.append(settings)